requirements.txt
fastapi
asyncpg
pydantic-settings
sqlmodel
alembic
passlib
pyjwt
aioredis
itsdangerous
celery
asgiref
langchain
langchain_community
